Moved the contents of the 'plutus-core-builtin' subproject to other subprojects
