+++ .hlint.yaml
@@ -1,6 +1,6 @@
 ---
 - functions:
-  - {name: unsafePerformIO, within: [PlutusPrelude, Language.PlutusCore.Generators.Internal.Entity, Language.PlutusCore.Builtin.Call, Language.PlutusCore.Builtin.Instances, Language.Plutus.CoreToPLC.Plugin]}
+  - {name: unsafePerformIO, within: [PlutusPrelude, Language.PlutusCore.Generators.Internal.Entity, Language.PlutusCore.Constant.Dynamic.Call, Language.PlutusCore.Constant.Dynamic.Emit, Language.PlutusCore.Constant.Dynamic.Instances, Language.Plutus.CoreToPLC.Plugin]}
   - {name: error, within: [Main, PlutusPrelude, Language.PlutusCore.StdLib.Meta, Evaluation.Constant.Success, Language.PlutusCore.Constant.Apply, Language.PlutusCore.Constant.Typed, Language.PlutusCore.Evaluation.CkMachine, Language.PlutusCore.TypeSynthesis, Language.PlutusCore.Generators.Internal.Entity, Language.PlutusCore.Generators.Internal.Utils, Language.PlutusCore.Constant.Make, Language.PlutusCore.Builtin.Instances, String, Language.PlutusCore.TH, Language.Plutus.CoreToPLC.Utils, Language.PlutusIR.Compiler.Datatype]}
   - {name: undefined, within: [Language.PlutusCore.Constant.Apply, Language.Plutus.Lift]}
   - {name: fromJust, within: [Language.Plutus.Lift]}
