+++ language-plutus-core/language-plutus-core.cabal
@@ -69,11 +69,13 @@ library
         Language.PlutusCore.Lexer.Type
         Language.PlutusCore.Parser
         Language.PlutusCore.Constant.Apply
-        Language.PlutusCore.Constant.DynamicInstances
+        Language.PlutusCore.Constant.Dynamic.Call
+        Language.PlutusCore.Constant.Dynamic.Emit
+        Language.PlutusCore.Constant.Dynamic.Instances
+        Language.PlutusCore.Constant.Dynamic.Pretty
         Language.PlutusCore.Constant.Function
         Language.PlutusCore.Constant.Make
         Language.PlutusCore.Constant.Name
-        Language.PlutusCore.Constant.PrettyDynamic
         Language.PlutusCore.Constant.Typed
         Language.PlutusCore.Check.Normal
         Language.PlutusCore.Pretty.Classic
