+++ language-plutus-core/src/Language/PlutusCore/Constant.hs
@@ -4,10 +4,12 @@ module Language.PlutusCore.Constant
     ( module Export
     ) where
 
-import           Language.PlutusCore.Constant.Apply            as Export
-import           Language.PlutusCore.Constant.DynamicInstances as Export ()
-import           Language.PlutusCore.Constant.Function         as Export
-import           Language.PlutusCore.Constant.Make             as Export
-import           Language.PlutusCore.Constant.Name             as Export
-import           Language.PlutusCore.Constant.PrettyDynamic    as Export
-import           Language.PlutusCore.Constant.Typed            as Export
+import           Language.PlutusCore.Constant.Apply             as Export
+import           Language.PlutusCore.Constant.Dynamic.Call      as Export
+import           Language.PlutusCore.Constant.Dynamic.Emit      as Export
+import           Language.PlutusCore.Constant.Dynamic.Instances as Export ()
+import           Language.PlutusCore.Constant.Dynamic.Pretty    as Export
+import           Language.PlutusCore.Constant.Function          as Export
+import           Language.PlutusCore.Constant.Make              as Export
+import           Language.PlutusCore.Constant.Name              as Export
+import           Language.PlutusCore.Constant.Typed             as Export
