+++ language-plutus-core/src/Language/PlutusCore/Constant/Dynamic/Call.hs
@@ -4,13 +4,15 @@
 {-# LANGUAGE RankNTypes       #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Language.PlutusCore.Builtin.Call
+module Language.PlutusCore.Constant.Dynamic.Call
     ( dynamicCallAssign
     , dynamicCall
     ) where
 
-import           Language.PlutusCore
-import           Language.PlutusCore.Constant
+import           Language.PlutusCore.Constant.Make
+import           Language.PlutusCore.Constant.Typed
+import           Language.PlutusCore.Lexer.Type     hiding (name)
+import           Language.PlutusCore.Type
 
 import           System.IO.Unsafe
