+++ language-plutus-core/src/Language/PlutusCore/Constant/Make.hs
@@ -26,8 +26,8 @@ module Language.PlutusCore.Constant.Make
     , makeBuiltinNOCHECK
     ) where
 
+import           Language.PlutusCore.Constant.Dynamic.Pretty
 import           Language.PlutusCore.Constant.Function
-import           Language.PlutusCore.Constant.PrettyDynamic
 import           Language.PlutusCore.Constant.Typed
 import           Language.PlutusCore.MkPlc
 import           Language.PlutusCore.Name
@@ -36,8 +36,8 @@ import           Language.PlutusCore.StdLib.Data.Bool
 import           Language.PlutusCore.Type
 import           PlutusPrelude
 
-import           Data.Bits                                  (bit)
-import qualified Data.ByteString.Lazy                       as BSL
+import           Data.Bits                                   (bit)
+import qualified Data.ByteString.Lazy                        as BSL
 import           Data.Maybe
 
 -- | Lift a 'BuiltinName' to 'Term'.
