+++ language-plutus-core/stdlib/Language/PlutusCore/StdLib/Data/List.hs
@@ -13,7 +13,7 @@ module Language.PlutusCore.StdLib.Data.List
     , getBuiltinProduct
     ) where
 
-import           Language.PlutusCore.Constant             (makeDynBuiltinInt)
+import           Language.PlutusCore.Constant.Make        (makeDynBuiltinInt)
 import           Language.PlutusCore.MkPlc
 import           Language.PlutusCore.Name
 import           Language.PlutusCore.Quote
