--- plutus-core-builtin/README.md
@@ -1,1 +1,0 @@
-# plutus-core-builtin
