--- plutus-core-builtin/shell.nix
@@ -1,1 +1,0 @@
-(import ../. {}).localPackages.plutus-core-builtin.env
