--- plutus-core-builtin/src/Language/PlutusCore/Builtin.hs
@@ -1,7 +1,0 @@
-module Language.PlutusCore.Builtin
-    ( module Export
-    ) where
-
-import           Language.PlutusCore.Builtin.Call      as Export
-import           Language.PlutusCore.Builtin.Common    as Export
-import           Language.PlutusCore.Builtin.Instances as Export ()
