+++ plutus-core-interpreter/plutus-core-interpreter.cabal
@@ -64,6 +64,9 @@ test-suite plutus-core-interpreter-test
     hs-source-dirs: test
     other-modules:
         CekMachine
+        Common
+        Factorial
+        String
     default-language: Haskell2010
     other-extensions: OverloadedStrings
     ghc-options: -threaded -rtsopts -with-rtsopts=-N -Wall
