+++ plutus-core-interpreter/test/Main.hs
@@ -1,6 +1,8 @@
 module Main (main) where
 
-import           CekMachine
+import           CekMachine (test_evaluateCek)
+import           Factorial  (test_dynamicFactorial)
+import           String     (test_dynamicStrings)
 
 import           Test.Tasty
 
@@ -8,6 +10,8 @@ test_Cek :: TestTree
 test_Cek =
     testGroup "Cek"
         [ test_evaluateCek
+        , test_dynamicFactorial
+        , test_dynamicStrings
         ]
 
 main :: IO ()
